+++ linkerd/app/outbound/src/http/logical.rs
@@ -1,5 +1,5 @@
 use super::{CanonicalDstHeader, Concrete, Endpoint, Logical};
-use crate::{resolve, stack_labels, target, Outbound};
+use crate::{endpoint, resolve, stack_labels, Outbound};
 use linkerd_app_core::{
     classify, config, profiles,
     proxy::{
@@ -66,10 +66,7 @@ impl<E> Outbound<E> {
             .check_service::<ConcreteAddr>()
             .push_request_filter(|c: Concrete| Ok::<_, Never>(c.resolve))
             .push(svc::layer::mk(move |inner| {
-                map_endpoint::Resolve::new(
-                    target::EndpointFromMetadata { identity_disabled },
-                    inner,
-                )
+                map_endpoint::Resolve::new(endpoint::FromMetadata { identity_disabled }, inner)
             }))
             .check_service::<Concrete>()
             .into_inner();
