+++ linkerd/app/outbound/src/tcp/mod.rs
@@ -5,14 +5,13 @@ pub mod opaque_transport;
 mod tests;
 
 pub use self::connect::Connect;
-use crate::target;
 pub use linkerd_app_core::proxy::tcp::Forward;
 use linkerd_app_core::{svc::Param, transport::OrigDstAddr, transport_header::SessionProtocol};
 
-pub type Accept = target::Accept<()>;
-pub type Logical = target::Logical<()>;
-pub type Concrete = target::Concrete<()>;
-pub type Endpoint = target::Endpoint<()>;
+pub type Accept = crate::Accept<()>;
+pub type Logical = crate::logical::Logical<()>;
+pub type Concrete = crate::logical::Concrete<()>;
+pub type Endpoint = crate::endpoint::Endpoint<()>;
 
 impl From<OrigDstAddr> for Accept {
     fn from(orig_dst: OrigDstAddr) -> Self {
@@ -23,7 +22,7 @@ impl From<OrigDstAddr> for Accept {
     }
 }
 
-impl<P> From<(P, Accept)> for target::Accept<P> {
+impl<P> From<(P, Accept)> for crate::Accept<P> {
     fn from((protocol, Accept { orig_dst, .. }): (P, Accept)) -> Self {
         Self { orig_dst, protocol }
     }
